fix(urlopy): modul urlopow zwracal 500 przy kazdym otwarciu
Na produkcji 36 razy w tygodniu, siedmiu z dziewieciu pracownikow nie moglo zobaczyc wniosku, salda ani kalendarza urlopowego.

Przyczyna: rzutowanie (int)ScopeLevel stalo WEWNATRZ zapytania LINQ, wiec EF tlumaczyl je na SQL 'scope_level::int'. Kolumna trzyma nazwy poziomow ('Organization', 'Department', 'Own'), wiec Postgres odrzucal zapytanie bledem 22P02. Konfiguracja EF ma poprawnie HasConversion<string>() - psulo to dopiero rzutowanie w projekcji.

Ten sam blad byl w dwoch miejscach: EmployeeScopeResolver (ujawnil sie w urlopach) i DataScopeService (jeszcze sie nie ujawnil, bo nikt nie wszedl w procesy).

Testy: nowy DataScopeSqlTranslationTests dziala na PRAWDZIWYM Postgresie, bo istniejace testy uzywaja dostawcy in-memory i z zasady nie moga wykryc bledu tlumaczenia na SQL - dlatego przepuscily to na produkcje. Sprawdzone, ze po cofnieciu poprawki test odtwarza dokladnie ten sam blad 22P02.

Trzeci test pilnuje zgodnosci wartosci obu typow wyliczeniowych, bo kod mapuje jeden na drugi przez int - cichy rozjazd nadalby lub odebral dostep do cudzych danych.

+++ src/WorkBase.Infrastructure/Auth/EmployeeScopeResolver.cs
@@ -75,13 +75,15 @@ public sealed class EmployeeScopeResolver(WorkBaseDbContext dbContext, IMemoryCa
             .ToListAsync(ct);
         if (roleIds.Count == 0) return DataScopeLevelValue.Team;
 
+        // Rzutowanie na int MUSI zostac poza zapytaniem: w projekcji EF tlumaczy je na
+        // scope_level::int, a kolumna trzyma nazwy poziomow i baza odrzuca zapytanie (22P02).
         var levels = await dbContext.Set<DataScope>()
             .Where(scope => scope.TenantId == tenantId && scope.Module == module && roleIds.Contains(scope.RoleId))
-            .Select(scope => (int)scope.ScopeLevel)
+            .Select(scope => scope.ScopeLevel)
             .ToListAsync(ct);
 
         // Tenanty sprzed wprowadzenia zakresów nie mają tych wierszy — nie odbieramy wtedy dostępu,
         // który daje samo uprawnienie zespołowe.
-        return levels.Count == 0 ? DataScopeLevelValue.Team : (DataScopeLevelValue)levels.Max();
+        return levels.Count == 0 ? DataScopeLevelValue.Team : (DataScopeLevelValue)(int)levels.Max();
     }
 }
